pyproject.toml
[tool.poetry]
name = "NerdBot"
version = "0.1.0"
description = "A discord bot that does many things"
authors = ["sulav maharjan <sulavmaharjan63@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
py-cord = "^1.7.3"
httpx = "^0.22.0"
beautifulsoup4 = "^4.10.0"
python-twitter = "^3.5"
python-decouple = "^3.5"
motor = "^2.5.1"
dnspython = "^2.2.0"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
